:sunrise: Initial commit
This commit creates the UNICEF Data Science & A.I. Toolkit, created for
the UNICEF Venture Fund. The project is initialized with a Hugo static
site built using the UNICEF Inventory theme. It comes prepared working
with a continuous integration pipeline for testing changes and deploying
them to production in GitHub Pages.

The project assumes a default license of CC-BY-SA-4.0.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>